pyproject.toml
[project]
name = "pd-graphiti-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "graphiti-core>=0.18.2",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.6.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "tenacity>=9.0.0",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
]
